Reduce top-level package dependencies
This commit extracts interfaces for ClientRSocketFactory,
ServerRSocketFactory, and ConnectionSetupPayload, and moves the
implementations into a sub-package with no references to it from the
top-level package.

This removes a large package cycle with io.rsocket at the center
of everything since it currently contains both high level
abstractions that (accessed by everything), as well as core protocol
implementations classes (accessing everything).

The refactoring is functionally neutral and backwards compatible.
